requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langgraph==0.0.26
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Database Drivers
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# NebulaGraph
nebula3-python==3.8.3

# Weaviate
weaviate-client==3.25.3

# Object Storage
boto3==1.34.10
minio==7.2.0

# Redis
redis==5.0.1
hiredis==2.3.2

# OCR & Document Processing
pytesseract==0.3.10
paddleocr==2.7.0.3
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
opencv-python==4.8.1.78

# Document Layout Analysis
transformers==4.36.0
torch==2.1.1
layoutparser==0.3.4

# NLP & Embeddings
sentence-transformers==2.2.2
spacy==3.7.2
openai==1.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
phonenumbers==8.13.27

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
langsmith==0.0.71

# Data Validation
email-validator==2.1.0
validators==0.22.0

# Task Queue (Optional)
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==21.0.0

# Code Quality
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2
